% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recdata.R
\docType{data}
\name{recdata}
\alias{recdata}
\title{Recruitment and survey index data}
\format{
Data frame containing 14 columns:
\tabular{ll}{
  \code{yearclass} \tab the yearclass\cr
  \code{recruitment} \tab the recruiment (age 3) for that yearclass\cr
  \code{NT1} \tab The age 1 survey index from 'NT' survey\cr
  \code{NT2} \tab The age 2 survey index from 'NT' survey\cr
  \code{...} \tab and so on
}
}
\usage{
recdata
}
\description{
data.frame containing recruitment (age 3) and survey indices from several
surveys over ages 1 to 3
}
\seealso{
\code{\link{rct3}}  run a calibrated regression to predict rectruitment.

\code{\link{rct3-package}} gives an overview of the package.
}
